+++ src/runtime-rs/crates/hypervisor/src/ch/inner_device.rs
@@ -5,11 +5,11 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use super::inner::CloudHypervisorInner;
+use crate::device::pci_path::PciPath;
 use crate::device::DeviceType;
 use crate::HybridVsockDevice;
 use crate::NetworkConfig;
 use crate::NetworkDevice;
-use crate::PciPath;
 use crate::ShareFsConfig;
 use crate::ShareFsDevice;
 use crate::VfioDevice;
