+++ src/runtime-rs/crates/hypervisor/src/device/mod.rs
@@ -17,6 +17,7 @@ use async_trait::async_trait;
 
 pub mod device_manager;
 pub mod driver;
+pub mod pci_path;
 pub mod util;
 
 #[derive(Debug)]
